@@ -1,9 +1,5 @@
+import { createApp } from "vue";
+import App from "./App.vue";
 import "./scss/main.css";
 
-import Vue from "vue";
-import App from "./App.vue";
-
-Vue.config.productionTip = false;
-new Vue({
-  render: (h) => h(App),
-}).$mount("#app");
+createApp(App).mount("#app");
